Revert "Unsupervised learning experiment (#15)"
This reverts commit 203fe0be2c6d0ed421967179e4f722707257a32b.
