+++ safe_opax/configs/agent/la_mbda.yaml
@@ -1,5 +1,5 @@
 defaults:
-  - penalizer: lbsgd
+  - penalizer: lagrangian
 
 name: lambda
 replay_buffer:
