+++ safe_opax/configs/environment/safe_adaptation_gym.yaml
@@ -1,5 +1,4 @@
 safe_adaptation_gym:
-  task: null
   robot_name: "point"
   image_observation:
     enabled: true
