+++ safe_opax/configs/experiment/safe_sparse_cartpole.yaml
@@ -1,6 +1,7 @@
 # @package _global_
 defaults:
   - override /environment: dm_cartpole
+  - override /agent/penalizer: lbsgd
 
 environment:
   dm_cartpole:
