--- safe_opax/configs/experiment/unsupervised.yaml
@@ -1,14 +1,0 @@
-# @package _global_
-defaults:
-  - override /environment: safe_adaptation_gym
-
-training:
-  trainer: unsupervised
-  epochs: 100
-  safe: true
-  action_repeat: 2
-  exploration_steps: 250000
-
-agent:
-  exploration_strategy: opax
-  exploration_steps: 250000
